Prevent err when dir doesnt exist

+++ scripts/startup/omnisharp-setup.js
@@ -22,22 +22,24 @@ rmdir(dstPath, () => {
 
 // https://gist.github.com/tkihira/2367067
 function rmdir(dir, cb) {
-	var list = fs.readdirSync(dir);
-	for(var i = 0; i < list.length; i++) {
-		var filename = path.join(dir, list[i]);
-		var stat = fs.statSync(filename);
-		
-		if(filename == "." || filename == "..") {
-			// pass these files
-		} else if(stat.isDirectory()) {
-			// rmdir recursively
-			rmdir(filename);
-		} else {
-			// rm fiilename
-			fs.unlinkSync(filename);
+	if (fs.existsSync(dir)) {
+		var list = fs.readdirSync(dir);
+		for(var i = 0; i < list.length; i++) {
+			var filename = path.join(dir, list[i]);
+			var stat = fs.statSync(filename);
+			
+			if(filename == "." || filename == "..") {
+				// pass these files
+			} else if(stat.isDirectory()) {
+				// rmdir recursively
+				rmdir(filename);
+			} else {
+				// rm fiilename
+				fs.unlinkSync(filename);
+			}
 		}
+		fs.rmdirSync(dir);
 	}
-	fs.rmdirSync(dir);
     if (cb) {
         cb();
     }
